rework auth module
