+++ src/authorization/authorization.module.ts
@@ -6,10 +6,17 @@ import { User } from '../user/user.entity';
 import { UserService } from '../user/user.service';
 import { JwtService } from '../jwt/jwt.service';
 import { MailService } from '../mail/mail.service';
+import { BcryptService } from '../bcrypt/bcrypt.service';
 
 @Module({
   controllers: [AuthorizationController],
-  providers: [AuthorizationService, UserService, JwtService, MailService],
+  providers: [
+    AuthorizationService,
+    UserService,
+    JwtService,
+    MailService,
+    BcryptService,
+  ],
   imports: [TypeOrmModule.forFeature([User])],
 })
 export class AuthorizationModule {}
