+++ src/user/user.service.ts
@@ -30,17 +30,8 @@ export class UserService {
     return { result, token };
   }
 
-  async confirmEmail({ token }) {
-    const { result, error } = this.jwtService.decodeToken(
-      token,
-      process.env.JWT_CONF_KEY,
-    );
-
-    if (error) {
-      throw new HttpException('Invalid token', 401);
-    }
-
-    const user = await this.getUserByEmail(result.email);
+  async checkConfirmEmail(email: string) {
+    const user = await this.getUserByEmail(email);
 
     if (!user) {
       throw new HttpException('User with this email not found', 400);
@@ -56,7 +47,7 @@ export class UserService {
         .where({ id: user.id })
         .execute();
       throw new HttpException(
-        'Sorry time to confirm your mail was expired',
+        'Sorry time to confirm your mail was expired, try to resign up',
         401,
       );
     }
@@ -73,90 +64,7 @@ export class UserService {
     return this.userRepository.findOne({ where: { email } });
   }
 
-  async signIn(data) {
-    const result = await this.getUserByEmail(data.email);
-
-    if (!result) {
-      throw new HttpException('Login or password is wrong', 403);
-    }
-
-    if (
-      !(await this.bcryptService.comparePassword(
-        data.password,
-        result.password,
-      ))
-    ) {
-      throw new HttpException('Login or password is wrong', 403);
-    }
-
-    if (!result.activated_at) {
-      throw new HttpException('Please verify your account', 403);
-    }
-
-    const token = this.jwtService.generateToken(
-      result,
-      process.env.JWT_ACCESS_KEY,
-    );
-
-    return { data: { message: 'Successfully Entered', token } };
-  }
-
-  async forgotPassword({ email }) {
-    const result = await this.getUserByEmail(email);
-
-    if (!result) {
-      throw new HttpException('User with this email not found', 400);
-    }
-
-    const token = this.jwtService.generateToken(
-      result,
-      process.env.JWT_CONF_KEY,
-    );
-
-    await this.mailService.sendUserMail({
-      email: result.email,
-      token,
-      text: EmailTextEnum.FORGOT_PASS_EMAIL,
-      subject: EmailSubjectEnum.FORGOT_PASS_EMAIL,
-      path: mail.PATH_FORGOT_PASS,
-    });
-
-    return { data: 'Check your mail', status: 200 };
-  }
-
-  async forgotPassMail({ token }) {
-    const { result, error } = this.jwtService.decodeToken(
-      token,
-      process.env.JWT_CONF_KEY,
-    );
-
-    if (error) {
-      throw new HttpException('Invalid token', 401);
-    }
-
-    const user = await this.getUserByEmail(result.email);
-
-    if (!user) {
-      throw new HttpException('User with this email not found', 400);
-    }
-  }
-
-  async changePassword({ password }, { token }) {
-    const { result, error } = this.jwtService.decodeToken(
-      token,
-      process.env.JWT_CONF_KEY,
-    );
-
-    if (error) {
-      throw new HttpException('Invalid token', 401);
-    }
-
-    const user = await this.getUserByEmail(result.email);
-
-    if (!user) {
-      throw new HttpException('User with this email not found', 400);
-    }
-
+  async changePassword(user, password) {
     user.password = await this.bcryptService.hashPassword(password);
     await this.userRepository
       .createQueryBuilder()
@@ -164,7 +72,5 @@ export class UserService {
       .set({ password: user.password })
       .where({ id: user.id })
       .execute();
-
-    return 'Your password successfully updated';
   }
 }
